backend/pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["brs_backend"]

[project]
name = "brs_backend"
version = "0.1.0"
description = "Backend service for BRS prototype"
authors = [{name = "BRS Team", email = "team@brs.example.com"}]
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.32.0",
    "sqlalchemy>=2.0.36",
    "psycopg2-binary>=2.9.9",
    "pydantic>=2.10.0",
    "smolagents[openai]>=1.21.2",
    "python-multipart>=0.0.12",
    "python-dotenv>=1.0.1",
    "PyJWT>=2.8.0",
    "passlib[bcrypt]>=1.7.4",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "pytest-asyncio>=0.21.0",
    "black>=23.0.0",
    "ruff>=0.1.0",
]

[dependency-groups]
dev = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.24.0",
    "httpx>=0.27.0",
    "black>=24.0.0",
    "ruff>=0.7.0",
]

[tool.black]
line-length = 88
target-version = ['py311']

[tool.ruff]
line-length = 88
target-version = "py311"
select = ["E", "F", "W", "I", "N", "B", "C4", "SIM", "TCH"]
ignore = ["E501"]  # Line too long (handled by black)

[tool.ruff.isort]
known-first-party = ["brs_backend"]


